claude-dc-implementation/computeruse/computer_use_demo_grok/requirements.txt
anthropic==0.50.0
pydantic==2.5.2
streamlit==1.31.0
nest-asyncio==1.5.8
python-dotenv==1.0.0
asyncio==3.4.3
pyautogui==0.9.54
pillow==10.2.0